Add files via upload
Here was the corpus that I had been building manually. It is only the first 50 or so from the Full URL list. My thought was just to have something to work with for the scope of the class with the goal of reading this data in more efficiently for the scope of the project.